+++ __tests__/test-error-messages.js
@@ -381,12 +381,7 @@ test.each(ABIVersions)('Test SDK Errors > 2000 (ABI v%i)', async (abiVersion) =>
             keyPair: wrongKeys,
         });
     } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2001);
-        expect(error.data)
-            .toBeNull();
+        expect(error.code).toEqual(2001);
     }
 
 
@@ -401,12 +396,7 @@ test.each(ABIVersions)('Test SDK Errors > 2000 (ABI v%i)', async (abiVersion) =>
             keyPair: wrongKeys,
         });
     } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2002);
-        expect(error.data)
-            .toBeNull();
+        expect(error.code).toEqual(2002);
     }
 
     wrongKeys = {
@@ -420,32 +410,17 @@ test.each(ABIVersions)('Test SDK Errors > 2000 (ABI v%i)', async (abiVersion) =>
             keyPair: wrongKeys,
         });
     } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2003);
-        expect(error.data)
-            .toBeNull();
+        expect(error.code).toEqual(2003);
     }
     try {
         await crypto.factorize('');
     } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2007);
-        expect(error.data)
-            .toBeNull();
+        expect(error.code).toEqual(2007);
     }
     try {
         await crypto.factorize('       ');
     } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2007);
-        expect(error.data)
-            .toBeNull();
+        expect(error.code).toEqual(2007);
     }
     try {
         await crypto.mnemonicFromEntropy({
@@ -453,12 +428,7 @@ test.each(ABIVersions)('Test SDK Errors > 2000 (ABI v%i)', async (abiVersion) =>
             dictionary: TONMnemonicDictionary.ENGLISH,
         });
     } catch (error) {
-        expect(error.source)
-            .toEqual('client');
-        expect(error.code)
-            .toEqual(2016);
-        expect(error.data)
-            .toBeNull();
+        expect(error.code).toEqual(2016);
     }
 
     await expectError(2017, 'client', async () => {